Move AidResource in a separate module
